refactor: renames files and icon components for readability purposes
